+++ src/router/index.js
@@ -3,6 +3,7 @@ import Router from 'vue-router';
 // import Hello from 'components/Hello'
 // 直接使用 components路径
 
+import Index from '../components/index/Index.vue';
 import Home from '../components/home/Home.vue';
 import Member from '../components/member/Member.vue';
 import Classify from '../components/classify/Classify.vue';
@@ -11,69 +12,79 @@ import My from '../components/my/My.vue';
 
 Vue.use(Router);
 
-const scrollBehavior = (to, from, savedPosition) => {
-  let data = JSON.parse(window.localStorage.getItem(to.name));
-  if (savedPosition) {
-    return savedPosition;
-  } else {
-    let y = 0;
-    if (data) {
-      y = data.y;
-    }
-    return {
-      x: 0,
-      y: y
-    };
-  }
-};
+// const scrollBehavior = (to, from, savedPosition) => {
+//   let data = JSON.parse(window.localStorage.getItem(to.name));
+//   if (savedPosition) {
+//     return savedPosition;
+//   } else {
+//     let y = 0;
+//     if (data) {
+//       y = data.y;
+//     }
+//     return {
+//       x: 0,
+//       y: y
+//     };
+//   }
+// };
 
 export default new Router({
   mode: 'history',
   routes: [
     {
       path: '/',
-      redirect: '/home'
+      redirect: '/index'
     },
     {
-      path: '/home',
-      name: 'home',
-      component: Home,
-      meta: {
-        saved: true
-      }
-    },
-    {
-      path: '/member',
-      name: 'member',
-      component: Member,
-      meta: {
-        saved: true
-      }
-    },
-    {
-      path: '/classify',
-      name: 'classify',
-      component: Classify,
-      meta: {
-        saved: true
-      }
-    },
-    {
-      path: '/shoppingcar',
-      name: 'shoppingcar',
-      component: ShoppingCar,
-      meta: {
-        saved: true
-      }
-    },
-    {
-      path: '/my',
-      name: 'my',
-      component: My,
-      meta: {
-        saved: true
-      }
+      path: '/index',
+      name: 'index',
+      component: Index,
+      children: [
+        {
+          path: '',
+          component: Home
+        },
+        {
+          path: 'home',
+          name: 'home',
+          component: Home,
+          meta: {
+            saved: true
+          }
+        },
+        {
+          path: 'member',
+          name: 'member',
+          component: Member,
+          meta: {
+            saved: true
+          }
+        },
+        {
+          path: 'classify',
+          name: 'classify',
+          component: Classify,
+          meta: {
+            saved: true
+          }
+        },
+        {
+          path: 'shoppingcar',
+          name: 'shoppingcar',
+          component: ShoppingCar,
+          meta: {
+            saved: true
+          }
+        },
+        {
+          path: 'my',
+          name: 'my',
+          component: My,
+          meta: {
+            saved: true
+          }
+        }
+      ]
     }
-  ],
-  scrollBehavior
+  ]
 });
